requirements.txt
pytest ~= 7.2.0
sphinx ~= 5.3.0
sphinx-rtd-theme ~= 1.1.1
opencv-python ~= 4.6.0.66
numpy ~= 1.24.0
scikit-image ~= 0.19.3
Pillow ~= 9.3.0